MovieDetailsPage

+++ app/src/main/java/com/test/bemoapplication/model/movielist/MovieResult.java
@@ -21,7 +21,7 @@ public class MovieResult {
     private Boolean video;
     @SerializedName("vote_average")
     @Expose
-    private Integer voteAverage;
+    private Float voteAverage;
     @SerializedName("title")
     @Expose
     private String title;
@@ -77,11 +77,11 @@ public class MovieResult {
         this.video = video;
     }
 
-    public Integer getVoteAverage() {
+    public Float getVoteAverage() {
         return voteAverage;
     }
 
-    public void setVoteAverage(Integer voteAverage) {
+    public void setVoteAverage(Float voteAverage) {
         this.voteAverage = voteAverage;
     }
